change default int type to i64
numpy defaults to 64-bit integers on most architectures. Also improves error messages when chaining fails to include the atom type.

+++ python/src/opendp/_lib.py
@@ -5,7 +5,7 @@ from typing import Optional, Any
 
 # list all acceptable alternative types for each default type
 ATOM_EQUIVALENCE_CLASSES = {
-    'i32': ['u8', 'u16', 'u32', 'u64', 'i8', 'i16', 'i32', 'i64', 'usize'],
+    'i64': ['u8', 'u16', 'u32', 'u64', 'i8', 'i16', 'i32', 'i64', 'usize'],
     'f64': ['f32', 'f64'],
     'bool': ['bool'],
     'AnyMeasurementPtr': ['AnyMeasurementPtr'],
